refactor: documentBuilder 의존성 변경
