+++ apps/blog/src/app/page.tsx
@@ -1,5 +1,5 @@
 import { groupPostsByYear } from "@/service/mdx";
-import { DocumentBuilder } from "@/service/mdx/post";
+import { DocumentBuilder } from "@/service/mdx";
 import { Header } from "@/widgets/Header";
 import { YearGroupPostList } from "@/widgets/posts/PostList";
