+++ apps/blog/src/service/mdx/index.ts
@@ -1,4 +1,5 @@
 export { groupPostsByYear } from "./groupPostsByYear";
 export { parseToc } from "./parseToc";
-export { getPostNavigation, allPosts, getPostsWithQuerys } from "./post";
+export { getPostNavigation, allPosts } from "./post";
 export { parseDateStringToDot } from "./parseDateStringToDot";
+export { DocumentBuilder } from "./post";
